Defeating boats and joker behaviour
If two boats of the same color are added to the deck of offered cards,
the game can now figure out if the player defeats the second one and
acts accordingly. I also added a joker behaviour so in case any player
is not able to defeat a boat and has to end the turn, all the players
with jokers get a coin for each joker they have.